Move logging devices on startup to controller.js. #45
