+++ lib/controller.js
@@ -27,6 +27,12 @@ class Controller {
             if (error) {
                 logger.error('Failed to start');
             } else {
+                // Log zigbee clients on startup.
+                const devices = this.zigbee.getAllClients();
+                logger.info(`Currently ${devices.length} devices are joined:`);
+                devices.forEach((device) => logger.info(this.getDeviceStartupLogMessage(device)));
+
+                // Connect to MQTT broker
                 const subscriptions = [
                     `${settings.get().mqtt.base_topic}/+/set`,
                     `${settings.get().mqtt.base_topic}/+/+/set`,
@@ -60,6 +66,21 @@ class Controller {
     stop(callback) {
         this.mqtt.disconnect();
         this.zigbee.stop(callback);
+    }
+
+    getDeviceStartupLogMessage(device) {
+        let friendlyName = 'unknown';
+        let friendlyDevice = {model: 'unkown', description: 'unknown'};
+
+        if (deviceMapping[device.modelId]) {
+            friendlyDevice = deviceMapping[device.modelId];
+        }
+
+        if (settings.getDevice(device.ieeeAddr)) {
+            friendlyName = settings.getDevice(device.ieeeAddr).friendly_name
+        }
+
+        return `${friendlyName} (${device.ieeeAddr}): ${friendlyDevice.model} - ${friendlyDevice.vendor} ${friendlyDevice.description}`;
     }
 
     handleZigbeeMessage(message) {
