+++ lib/zigbee.js
@@ -1,7 +1,6 @@
 const ZShepherd = require('zigbee-shepherd');
 const logger = require('./util/logger');
 const settings = require('./util/settings');
-const deviceMapping = require('./devices');
 const data = require('./util/data');
 
 const shepherdSettings = {
@@ -46,15 +45,9 @@ class Zigbee {
     }
 
     handleReady() {
-        logger.info('zigbee-shepherd ready');
-
-        const devices = this.getAllClients();
-
-        logger.info(`Currently ${devices.length} devices are joined:`);
-        devices.forEach((device) => logger.info(this.getDeviceLogMessage(device)));
-
         // Set all Xiaomi devices (manufId === 4151) to be online, so shepherd won't try
         // to query info from devices (which would fail because they go tosleep).
+        const devices = this.getAllClients();
         devices.forEach((device) => {
             if (device.manufId === 4151) {
                 this.shepherd.find(device.ieeeAddr, 1).getDevice().update({
@@ -63,6 +56,8 @@ class Zigbee {
                 });
             }
         });
+
+        logger.info('zigbee-shepherd ready');
     }
 
     permitJoin(permit) {
@@ -87,21 +82,6 @@ class Zigbee {
         if (this.onMessage) {
             this.onMessage(message);
         }
-    }
-
-    getDeviceLogMessage(device) {
-        let friendlyName = 'unknown';
-        let friendlyDevice = {model: 'unkown', description: 'unknown'};
-
-        if (deviceMapping[device.modelId]) {
-            friendlyDevice = deviceMapping[device.modelId];
-        }
-
-        if (settings.getDevice(device.ieeeAddr)) {
-            friendlyName = settings.getDevice(device.ieeeAddr).friendly_name
-        }
-
-        return `${friendlyName} (${device.ieeeAddr}): ${friendlyDevice.model} - ${friendlyDevice.vendor} ${friendlyDevice.description}`;
     }
 
     getDevice(deviceID) {
